requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.1.8
dotenv==0.9.9
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector==2.2.9
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
